requirements.txt
atomicwrites==1.3.0
attrs==19.1.0
Click==7.0
Flask==1.0.2
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
more-itertools==7.0.0
numpy==1.16.3
openreview-py==0.9.24
ortools==7.0.6546
pluggy==0.11.0
protobuf==3.7.1
py==1.8.0
pytest==4.4.1
six==1.12.0
Werkzeug==0.15.2
